# old_code/src/docs/swagger.json
{"info":{"title":"Vehicles API","version":"1.0.0","description":""},"openapi":"3.0.1","paths":{"/api/vehicle":{"get":{"tags":["Vehicles"],"summary":"Return a list of vehicles.","parameters":[{"in":"query","name":"page","required":true,"schema":{"type":"number","minimum":1}},{"in":"query","name":"limit","required":true,"schema":{"type":"number","minimum":0}}],"responses":{"200":{"description":"Returns a list of vehicles.","content":{"application/json":{"example":{"page":1,"limit":10,"totalPages":10,"totalItems":20,"vehicles":[{"id":1,"brand":"Ford","model":"Fiesta","year":2000,"chassis":"9BWZZZ377VT004251","plate":"ABC-2000","renavam":"12345678901"},{"id":2,"brand":"Ford","model":"Fiesta","year":2001,"chassis":"9BWZZZ377VT004251","plate":"ABC-2001","renavam":"12345678901"},{"id":3,"brand":"Ford","model":"Fiesta","year":2002,"chassis":"9BWZZZ377VT004251","plate":"ABC-2002","renavam":"12345678901"},{"id":4,"brand":"Ford","model":"Fiesta","year":2003,"chassis":"9BWZZZ377VT004251","plate":"ABC-2003","renavam":"12345678901"},{"id":5,"brand":"Ford","model":"Fiesta","year":2004,"chassis":"9BWZZZ377VT004251","plate":"ABC-2004","renavam":"12345678901"},{"id":6,"brand":"Ford","model":"Fiesta","year":2005,"chassis":"9BWZZZ377VT004251","plate":"ABC-2005","renavam":"12345678901"},{"id":7,"brand":"Ford","model":"Fiesta","year":2006,"chassis":"9BWZZZ377VT004251","plate":"ABC-2006","renavam":"12345678901"},{"id":8,"brand":"Ford","model":"Fiesta","year":2007,"chassis":"9BWZZZ377VT004251","plate":"ABC-2007","renavam":"12345678901"},{"id":9,"brand":"Ford","model":"Fiesta","year":2008,"chassis":"9BWZZZ377VT004251","plate":"ABC-2008","renavam":"12345678901"},{"id":10,"brand":"Ford","model":"Fiesta","year":2009,"chassis":"9BWZZZ377VT004251","plate":"ABC-2009","renavam":"12345678901"}]}}}},"400":{"description":"Invalid query parameters. Check the page and limit parameters.","content":{"application/json":{"examples":{"Invalid query parameter.":{"value":{"message":"Invalid query parameters. Check the page and limit parameters."}}}}}},"405":{"description":"Method not allowed.","content":{"application/json":{"examples":{"Method not allowed":{"value":{"message":"Method not allowed."}}}}}},"500":{"description":"An error occurred in our server.","content":{"application/json":{"examples":{"Invalid query parameters":{"value":{"message":"An error occurred in our server."}}}}}}}},"post":{"tags":["Vehicles"],"summary":"Create a new vehicle.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"brand":{"type":"string"},"model":{"type":"string"},"year":{"example":2023,"type":"number"},"chassis":{"type":"string"},"plate":{"type":"string"},"renavam":{"type":"string"}},"required":["brand","model","year","chassis","plate","renavam"]}}}},"responses":{"201":{"description":"Creates and return a vehicle object.","content":{"application/json":{"example":{"vehicle":{"id":1,"brand":"Ford","model":"Fiesta","year":2023,"chassis":"9BWZZZ377VT004251","plate":"ABC-1234","renavam":"12345678901"}}}}},"400":{"description":"Invalid body request","content":{"application/json":{"examples":{"Vehicle already exists.":{"description":"Vehicle already exists.","value":{"message":"Vehicle already exists."}},"Invalid body request.":{"description":"Invalid body request. Check your request body.","value":{"message":"Invalid body request. Check your request body."}}}}}},"405":{"description":"Method not allowed.","content":{"application/json":{"examples":{"Method not allowed":{"value":{"message":"Method not allowed."}}}}}},"500":{"description":"An error occurred in our server.","content":{"application/json":{"examples":{"Invalid query parameters":{"value":{"message":"An error occurred in our server."}}}}}}}}},"/api/vehicle/{plate}":{"get":{"tags":["Vehicles"],"summary":"Return a vehicle.","parameters":[{"in":"path","name":"plate","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a vehicle object.","content":{"application/json":{"example":{"vehicle":{"id":1,"brand":"Ford","model":"Fiesta","year":2023,"chassis":"9BWZZZ377VT004251","plate":"ABC-1234","renavam":"12345678901"}}}}},"400":{"description":"Invalid url parameter. Check plate parameter.","content":{"application/json":{"examples":{"Invalid url parameter.":{"value":{"message":"Invalid url parameter. Check plate parameter."}}}}}},"405":{"description":"Method not allowed.","content":{"application/json":{"examples":{"Method not allowed":{"value":{"message":"Method not allowed."}}}}}},"500":{"description":"An error occurred in our server.","content":{"application/json":{"examples":{"Invalid query parameters":{"value":{"message":"An error occurred in our server."}}}}}}}},"put":{"tags":["Vehicles"],"summary":"Update a new vehicle.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"brand":{"type":"string"},"model":{"type":"string"},"year":{"example":2023,"type":"number"},"chassis":{"type":"string"},"renavam":{"type":"string"}}}}}},"parameters":[{"in":"path","name":"plate","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Creates and return a vehicle object.","content":{"application/json":{"example":{"vehicle":{"id":1,"brand":"Ford","model":"Fiesta","year":2023,"chassis":"9BWZZZ377VT004251","plate":"ABC-1234","renavam":"12345678901"}}}}},"400":{"description":"Invalid body request. Check your request body.","content":{"application/json":{"examples":{"Invalid query parameter.":{"value":{"message":"Invalid query parameters. Check the page and limit parameters."}},"Vehicle not found.":{"message":"Vehicle not found."}}}}},"405":{"description":"Method not allowed.","content":{"application/json":{"examples":{"Method not allowed":{"value":{"message":"Method not allowed."}}}}}},"500":{"description":"An error occurred in our server.","content":{"application/json":{"examples":{"Invalid query parameters":{"value":{"message":"An error occurred in our server."}}}}}}}},"delete":{"tags":["Vehicles"],"summary":"Delete a vehicle.","parameters":[{"in":"path","name":"plate","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Delete and return a vehicle object.","content":{"application/json":{"example":{"vehicle":{"id":1,"brand":"Ford","model":"Fiesta","year":2023,"chassis":"9BWZZZ377VT004251","plate":"ABC-1234","renavam":"12345678901"}}}}},"400":{"description":"Invalid url parameter. Check plate parameter.","content":{"application/json":{"examples":{"Invalid query parameter.":{"value":{"message":"Invalid url parameter. Check plate parameter."}},"Vehicle not found.":{"message":"Vehicle not found."}}}}},"405":{"description":"Method not allowed.","content":{"application/json":{"examples":{"Method not allowed":{"value":{"message":"Method not allowed."}}}}}},"500":{"description":"An error occurred in our server.","content":{"application/json":{"examples":{"Invalid query parameters":{"value":{"message":"An error occurred in our server."}}}}}}}}}}}